警告を回避

+++ Team16/Actor/Enemies/BomEnemy.cpp
@@ -50,7 +50,7 @@ void BomEnemy::draw(Renderer * renderer, Renderer3D* renderer3D)
 {
 	if (b_mType == Type::ENEMY)
 	{
-		renderer3D->draw3DTexture("enemyR2", Vector3(b_mPosittion.x, b_mPosittion.y, 0.0f), Vector2(0.0f, 0.0f), Vector2(64.0f, 64.0f), 96.0f, b_mAngle, 255, Vector2(0.5f, 0.5f), Vector3(255, mDamageHit, mDamageHit));
+		renderer3D->draw3DTexture("enemyR2", Vector3(b_mPosittion.x, b_mPosittion.y, 0.0f), Vector2(0.0f, 0.0f), Vector2(64.0f, 64.0f), 96.0f, b_mAngle, 255, Vector2(0.5f, 0.5f), Vector3((float)255, (float)mDamageHit, (float)mDamageHit));
 		if (b_mHp <= 0)
 		{
 			b_animCnt += 64.0f;
